ci: Add dependabot automation workflow (#73)
This commit adds a workflow to automate some tedious tasks that I currently have to handle whenever a new Dependabot pull request is opened.

The workflow will first check that there are no failed checks, other than the documentation check, to ensure failed checks aren't automatically merged. This is necessary since checks are not triggered for commits made using the GITHUB_TOKEN automatically created for a given workflow run. Otherwise, I would have been able to blindly update the docs and that check would pass on the next check run.

Once the checks have been validated, the documentation (README) is updated via terraform-docs which is required for every Dependabot commit (and is what I am looking to automate away).

Finally, the pull request is approved and auto-merge is enabled such that the Dependabot PR can be merged and no manual intervention is required.